Split graph navigator into two classes
