+++ src/JMS/Serializer/Serializer.php
@@ -164,7 +164,7 @@ class Serializer implements SerializerInterface
 
     private function visit(VisitorInterface $visitor, Context $context, $data, $format, array $type = null)
     {
-        $navigator = $this->navigatorFactory->getGraphNavigator();
+        $navigator = $this->navigatorFactory->getGraphNavigator($context instanceof DeserializationContext ? GraphNavigator::DIRECTION_DESERIALIZATION : GraphNavigator::DIRECTION_SERIALIZATION);
         $context->initialize(
             $format,
             $visitor,
